requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
dnspython>=2.6.0
openpyxl>=3.1.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
python-multipart>=0.0.9
httpx>=0.27.0
